Three big changes
Firstly, changed the bootstrap algorithm so that it conforms with
pROC and ROCr. The sampling is now of the entire dataset, with
replacement. Previously, i only sampled parts of the dataset at one
time.

Secondly, added parallelization through the "parallel" keyword to the
bootstrap functions.

Third, added one (1) test. Will add more!