Changed ROC thresholding strategy
Changed the ROC thresholding strategy.
It now depends on the unique predicted values, and computes intervals
between each predicted value like so:

(c(1,p) + c(p,0)) / 2

It is also possible to pass in your own intervals, either as a single
number to create a sequence between 0 and 1, or as a vector of unique
values. This last option was very useful for debugging.